+++ server/package.json
@@ -9,6 +9,7 @@
     "@google/generative-ai": "^0.5.0",
     "bcrypt": "^5.1.1",
     "body-parser": "^1.20.2",
+    "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "express-rate-limit": "^7.2.0",
