+++ server/src/lib/gemini.js
