+++ server/src/routes/aiRoutes.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const router = express.Router();
+const {getAi} = require("../controllers/aiController");
+
+router.get("/", getAi);
+
+module.exports = router;
