+++ server/src/routes/authRoutes.js
@@ -1,0 +1,11 @@
+const express = require("express");
+const router = express.Router();
+const {getRegister, getLogin} = require("../controllers/authController");
+
+// register
+router.post("/register", getRegister);
+
+// login
+router.post("/login", getLogin);
+
+module.exports = router;
