+++ server/src/routes/toggleRoutes.js
@@ -1,0 +1,12 @@
+const express = require("express");
+const router = express.Router();
+const {
+    getToggle,
+    updateToggle,
+} = require("../controllers/toggleController");
+
+router.put("/", updateToggle);
+
+router.get("/", getToggle);
+
+module.exports = router;
